chore: update Docker to use modern images
* update to use node:20-slim
* resolves warning that comes up with npm WARN
* update comment to clarify difference between
  using ci (clean install) over install which
  isn't necessarily faster (and can be in fact
  slower) and the context of why production.

This PR intentionally does not resolve some other
remaining Docker issues as there are remaining
issues to resolve including appropriate fonts
with Alpine image use.

This does resolve a chunk of the Dockerfile
updates. Many of these are exactly the same
and there may be room to centralize and
standardize rather than having separate files.